requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
aiosqlite>=0.20.0
anthropic>=0.42.0
google-genai>=1.0.0
apscheduler>=3.10.0
httpx>=0.28.0
python-multipart>=0.0.18
python-dotenv>=1.0.0
pydantic-settings>=2.7.0
bcrypt>=4.0.0
pymupdf>=1.24.0
openpyxl>=3.1.0
python-dateutil>=2.9.0
pycryptodome>=3.20.0
zeroconf>=0.131.0
